Fixed tracking of argument scope.
This properly catches unsupported "await" in arguments.

This fixes #730 issue on Github.

+++ src/njs_parser.h
@@ -26,7 +26,7 @@ struct njs_parser_scope_s {
     uint8_t                         arrow_function;
     uint8_t                         dest_disable;
     uint8_t                         async;
-    uint8_t                         in_args;
+    uint32_t                        in_args;
 };
